Adjusted the file structure
