+++ src/components/content-components/CardArea.js
